requirements.txt
# requirements.txt
streamlit
streamlit-option-menu
streamlit-image-comparison
pillow
numpy
# opencv-python # 로컬에서는 이것도 괜찮음
opencv-python-headless # 배포 시 권장 (GUI 의존성 제거)
mediapipe
requests
openai >= 1.0 # 최신 openai 라이브러리 버전 명시 # 추가 (이미지 비교 컴포넌트)